requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
redis>=4.6.0,<8.0.0
rq>=2.0.0,<3.0.0
pydantic==1.10.12
python-dotenv==1.0.0
